update ottrace propagator
