+++ propagator/ottrace/lib/opentelemetry/propagator/ottrace/text_map_propagator.rb
@@ -15,36 +15,38 @@ module OpenTelemetry
   module Propagator
     # Namespace for OpenTelemetry OTTrace propagation
     module OTTrace
-      # Extracts context from carriers using OTTrace header format
-      class TextMapExtractor
+      # Propagates context using OTTrace header format
+      class TextMapPropagator
         PADDING = '0' * 16
         VALID_TRACE_ID_REGEX = /^[0-9a-f]{32}$/.freeze
         VALID_SPAN_ID_REGEX = /^[0-9a-f]{16}$/.freeze
+        TRACE_ID_64_BIT_WIDTH = 64 / 4
+        TRACE_ID_HEADER = 'ot-tracer-traceid'
+        SPAN_ID_HEADER = 'ot-tracer-spanid'
+        SAMPLED_HEADER = 'ot-tracer-sampled'
+        BAGGAGE_HEADER_PREFIX = 'ot-baggage-'
+        FIELDS = [TRACE_ID_HEADER, SPAN_ID_HEADER, SAMPLED_HEADER].freeze
 
-        # Returns a new TextMapExtractor that extracts OTTrace context using the
-        # specified getter
-        #
-        # @param [optional Getter] default_getter The default getter used to read
-        #   headers from a carrier during extract. Defaults to a
-        #   {OpenTelemetry::Context:Propagation::TextMapGetter} instance.
-        # @return [TextMapExtractor]
-        def initialize(default_getter: Context::Propagation.text_map_getter)
-          @default_getter = default_getter
-        end
+        # https://github.com/open-telemetry/opentelemetry-specification/blob/14d123c121b6caa53bffd011292c42a181c9ca26/specification/context/api-propagators.md#textmap-propagator0
+        VALID_BAGGAGE_HEADER_NAME_CHARS = /^[\^_`a-zA-Z\-0-9!#$%&'*+.|~]+$/.freeze
+        INVALID_BAGGAGE_HEADER_VALUE_CHARS = /[^\t\u0020-\u007E\u0080-\u00FF]/.freeze
+
+        private_constant :PADDING, :VALID_TRACE_ID_REGEX, :VALID_SPAN_ID_REGEX, :TRACE_ID_64_BIT_WIDTH, :TRACE_ID_HEADER,
+                         :SPAN_ID_HEADER, :SAMPLED_HEADER, :BAGGAGE_HEADER_PREFIX, :FIELDS, :VALID_BAGGAGE_HEADER_NAME_CHARS,
+                         :INVALID_BAGGAGE_HEADER_VALUE_CHARS
 
         # Extract OTTrace context from the supplied carrier and set the active span
         # in the given context. The original context will be returned if OTTrace
         # cannot be extracted from the carrier.
         #
         # @param [Carrier] carrier The carrier to get the header from.
-        # @param [Context] context The context to be updated with extracted context
+        # @param [optional Context] context The context to be updated with extracted context
         # @param [optional Getter] getter If the optional getter is provided, it
         #   will be used to read the header from the carrier, otherwise the default
         #   getter will be used.
         # @return [Context] Updated context with active span derived from the header, or the original
         #   context if parsing fails.
-        def extract(carrier, context, getter = nil)
-          getter ||= default_getter
+        def extract(carrier, context: Context.current, getter: Context::Propagation.text_map_getter)
           trace_id = optionally_pad_trace_id(getter.get(carrier, TRACE_ID_HEADER))
           span_id = getter.get(carrier, SPAN_ID_HEADER)
           sampled = getter.get(carrier, SAMPLED_HEADER)
@@ -62,9 +64,22 @@ module OpenTelemetry
           Trace.context_with_span(span, parent_context: set_baggage(carrier: carrier, context: context, getter: getter))
         end
 
-        private
+        # @param [Object] carrier to update with context.
+        # @param [optional Context] context The active Context.
+        # @param [optional Setter] setter If the optional setter is provided, it
+        #   will be used to write context into the carrier, otherwise the default
+        #   setter will be used.
+        def inject(carrier, context: Context.current, setter: Context::Propagation.text_map_setter)
+          span_context = Trace.current_span(context).context
+          return unless span_context.valid?
 
-        attr_reader :default_getter
+          inject_span_context(span_context: span_context, carrier: carrier, setter: setter)
+          inject_baggage(context: context, carrier: carrier, setter: setter)
+
+          nil
+        end
+
+        private
 
         def valid?(trace_id:, span_id:)
           !(VALID_TRACE_ID_REGEX !~ trace_id || VALID_SPAN_ID_REGEX !~ span_id)
@@ -80,7 +95,7 @@ module OpenTelemetry
 
         def set_baggage(carrier:, context:, getter:)
           baggage.build(context: context) do |builder|
-            prefix = OTTrace::BAGGAGE_HEADER_PREFIX
+            prefix = BAGGAGE_HEADER_PREFIX
             getter.keys(carrier).each do |carrier_key|
               baggage_key = carrier_key.start_with?(prefix) && carrier_key[prefix.length..-1]
               next unless baggage_key
@@ -96,6 +111,22 @@ module OpenTelemetry
 
         def baggage
           OpenTelemetry.baggage
+        end
+
+        def inject_span_context(span_context:, carrier:, setter:)
+          setter.set(carrier, TRACE_ID_HEADER, span_context.hex_trace_id[TRACE_ID_64_BIT_WIDTH, TRACE_ID_64_BIT_WIDTH])
+          setter.set(carrier, SPAN_ID_HEADER, span_context.hex_span_id)
+          setter.set(carrier, SAMPLED_HEADER, span_context.trace_flags.sampled?.to_s)
+        end
+
+        def inject_baggage(context:, carrier:, setter:)
+          baggage.values(context: context)
+                 .select { |key, value| valid_baggage_entry?(key, value) }
+                 .each { |key, value| setter.set(carrier, "#{BAGGAGE_HEADER_PREFIX}#{key}", value) }
+        end
+
+        def valid_baggage_entry?(key, value)
+          VALID_BAGGAGE_HEADER_NAME_CHARS =~ key && INVALID_BAGGAGE_HEADER_VALUE_CHARS !~ value
         end
       end
     end
