+++ propagator/ottrace/test/opentelemetry/propagator/ottrace_test.rb
@@ -7,20 +7,11 @@
 require 'test_helper'
 
 describe OpenTelemetry::Propagator::OTTrace do
-  describe '#text_map_extractor' do
-    it 'returns an instance of TextMapExtractor' do
-      extractor = OpenTelemetry::Propagator::OTTrace.text_map_extractor
-      _(extractor).must_be_instance_of(
-        OpenTelemetry::Propagator::OTTrace::TextMapExtractor
-      )
-    end
-  end
-
-  describe '#text_map_injector' do
-    it 'returns an instance of TextMapInjector' do
-      extractor = OpenTelemetry::Propagator::OTTrace.text_map_injector
-      _(extractor).must_be_instance_of(
-        OpenTelemetry::Propagator::OTTrace::TextMapInjector
+  describe '#text_map_propagator' do
+    it 'returns an instance of TextMapPropagator' do
+      propagator = OpenTelemetry::Propagator::OTTrace.text_map_propagator
+      _(propagator).must_be_instance_of(
+        OpenTelemetry::Propagator::OTTrace::TextMapPropagator
       )
     end
   end
